+++ noodles-bed/examples/bed_write.rs
@@ -5,6 +5,7 @@
 use std::io;
 
 use noodles_bed as bed;
+use noodles_core::Position;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let stdout = io::stdout();
@@ -13,8 +14,8 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let record = bed::Record::<3>::builder()
         .set_reference_sequence_name("sq0")
-        .set_start_position(8)
-        .set_end_position(13)
+        .set_start_position(Position::try_from(8)?)
+        .set_end_position(Position::try_from(13)?)
         .build()?;
 
     writer.write_record(&record)?;