+++ noodles-bed/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * bed/record: Wrap start and end positions.
+
+    Positions are now 1-based, inclusive.
+
 ## 0.1.0 - 2022-02-17
 
   * bed: Initial release.